Add common flags to heartbeat in WakaTime extension
The code modifies the WakaTime extension to include common flags in the CliHeartbeat. It not only adds the common flags that are kept in flags holder of the WakaTime instance, but also introduces corresponding adjustments such as assigning common flags to a variable and adding them to the heartbeat object.