Refactor FlagHolder to Heartbeat class
The class FlagHolder is replaced with a class named Heartbeat. The refactoring involved replacing all instances of FlagHolder in WakaTime.cs with Heartbeat and removing validation methods from FlagHolder class that checks whether or not the heartbeat has all the required flags set. These methods are now in the Heartbeat class. The Heartbeat class is added as a new file in the codebase.